hw10/src/main/java/ru/otus/hw/controllers/AuthorRestController.java
package ru.otus.hw.controllers;

import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RestController;
import ru.otus.hw.dto.AuthorDTO;
import ru.otus.hw.services.AuthorService;

import java.util.List;

@RestController
@RequiredArgsConstructor
public class AuthorRestController {

    private final AuthorService authorService;

    @GetMapping("author")
    public ResponseEntity<List<AuthorDTO>> getAuthors() {
        var authors = authorService.findAll();
        return ResponseEntity.ok(authors);
    }

}
